Refactor out cardgame to separate example
